pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["UstymenkoOB <UstumenkoOB@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
tensorflow-io-gcs-filesystem = "0.31.0"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
fastapi-limiter = "0.1.5"
alembic = "^1.12.1"
pydantic = {extras = ["email"], version = "^2.6.4"}
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = "4.6.0"

cloudinary = "^1.36.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
qrcode = "^7.4.2"
asyncio = "^3.4.3"
pillow = "^10.2.0"
keras = "^3.0.5"
tensorflow = "^2.16.1"
numpy = "^1.26.4"
psycopg2 = "^2.9.9"
jinja2 = "^3.1.3"
fastapi-mail = "^1.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
sphinx = "^7.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
